requirements.txt
flask
flask_sqlalchemy
mongoengine
goose3
python-dateutil
scrapy
scrapy-rotating-proxies
celery
redis
apscheduler
python-twitter
pytrends
